Update FM and ADPCM emulation of MAME
